% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_model.R
\name{choose_best_model}
\alias{choose_best_model}
\title{Picks the best model from multiple initial conditions for the optimisation
to ensure the result is not from a local likelihood optima and is from the
global likelihood optima}
\usage{
choose_best_model(data_name, model)
}
\arguments{
\item{data_name}{Character string with the name of the island data set}

\item{model}{Character string with the name of the model fit to the data}
}
\value{
Data frame
}
\description{
Picks the best model from multiple initial conditions for the optimisation
to ensure the result is not from a local likelihood optima and is from the
global likelihood optima
}
\examples{
choose_best_model(data_name = "Canaries", model = "cr_dd")
}
