% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_results_as_df.R
\name{daisie_results_as_df}
\alias{daisie_results_as_df}
\title{Reads and stores the results from raw DAISIE model output into a data frame}
\usage{
daisie_results_as_df(islands)
}
\arguments{
\item{islands}{A string or vector of strings defining which island data sets
should be read and returned}
}
\value{
Data frame
}
\description{
Reads and stores the results from raw DAISIE model output into a data frame
}
\examples{
islands <- c("Canaries", "Comoros")
results_df <- daisie_results_as_df(islands)
}
