% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxedDAISIE-package.R
\docType{package}
\name{relaxedDAISIE-package}
\alias{relaxedDAISIE}
\alias{relaxedDAISIE-package}
\title{relaxedDAISIE: Package housing all the code to reproduce relaxed-rate DAISIE paper}
\description{
All of the scripts required to reproduce the analyses are kept in this package.
}
\keyword{internal}
