% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aic.R
\name{calc_aic}
\alias{calc_aic}
\title{Calculates Akaike's information criterion for a maximum likelihood output
of the DAISIE model}
\usage{
calc_aic(results, daisie_data)
}
\arguments{
\item{results}{The maximum likelihood parameter estimates and associated log
likelihood and free parameters (degrees of freedom)}

\item{daisie_data}{The data the DAISIE model was fit to}
}
\value{
A numeric
}
\description{
Calculates Akaike's information criterion for a maximum likelihood output
of the DAISIE model
}
\examples{
stub
}
