% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aic.R
\name{calc_aicc}
\alias{calc_aicc}
\title{Calculates the bias-corrected (second-order) Akaike's information criterion
for a maximum likelihood output of the DAISIE model}
\usage{
calc_aicc(results, daisie_data)
}
\arguments{
\item{results}{The maximum likelihood parameter estimates and associated log
likelihood and free parameters (degrees of freedom)}

\item{daisie_data}{The data the DAISIE model was fit to}
}
\value{
A numeric
}
\description{
Calculates the bias-corrected (second-order) Akaike's information criterion
for a maximum likelihood output of the DAISIE model
}
\examples{
stub
}
