% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr_cr_diff.R
\name{calc_rr_cr_diff}
\alias{calc_rr_cr_diff}
\title{Calculates the difference in likelihood for a clade (single lineage or
radiation) across different standard deviations of gamma distribution of
the relaxed-rate DAISIE model.}
\usage{
calc_rr_cr_diff(island_clade, relaxed_par, sd, par_upper_bound)
}
\arguments{
\item{island_clade}{A list with information on a single island clade,
including clade name, branching times (including island age and colonisation
time), stac (status of colonist), number of missing species and type of
clade (either type 1 or 2)}

\item{relaxed_par}{A string defining which parameter to relaxed. Options are:
"cladogenesis", "extinction", "carrying_capacity", "immigration" or
"anagenesis"}

\item{sd}{A vector of numerics with the standard deviation to be used to
check the difference between the original DAISIE model and the relaxed-rate
DAISIE model}

\item{par_upper_bound}{A numeric which sets the upper bound to the
integration of the likelihood, can be infinite or finite.}
}
\value{
A list of two numeric vectors
}
\description{
Calculates the difference in likelihood for a clade (single lineage or
radiation) across different standard deviations of gamma distribution of
the relaxed-rate DAISIE model.
}
