% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxedDAISIE-package.R
\docType{package}
\name{relaxedDAISIE-package}
\alias{relaxedDAISIE}
\alias{relaxedDAISIE-package}
\title{relaxedDAISIE: Reproduce Relaxed-rate DAISIE Paper}
\description{
All of the scripts required to reproduce the analyses and plots of the relaxed-rate DAISIE paper are kept in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshwlambert/relaxedDAISIE}
  \item Report bugs at \url{https://github.com/joshwlambert/relaxedDAISIE/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{j.w.l.lambert@rug.nl} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID})

}
\keyword{internal}
